LayoutTests/fast/images/text-recognition/ios/show-data-detector-context-menu-expected.txt

PASS result?.contextMenu?.items?.includes('Open Link') is true
PASS successfullyParsed is true

TEST COMPLETE
